test: 新增事件中心单元测试，转移 utils 的测试为单测
