+++ src/__tests__/unit/mocks/event.ts
@@ -1,0 +1,3 @@
+export const MOCK_EVENT_NAME = 'micro-app-event'
+
+export const MOCK_EVENT_DATA = { data: 'event data' }
